use constants, refactor hardcode, fix responsive touch/mousewheel events (e.g. scroll) not working on svg elements
