+++ src/components/Common/LandingPage/less/logo.less
@@ -69,6 +69,7 @@
     top: 0;
     left: 0;
     overflow: hidden;
+    pointer-events: none;
   }
   width: 100%;
   height: 200px;
