+++ src/constants/hardcode/landing/index.js
